feat(141): searchFilterChips 화면 변경 시 초기화되도록 수정
