+++ src/components/search/types.ts
@@ -6,4 +6,15 @@ export type SearchModalProps = {
 	setDisctrictSelectorOpen: Dispatch<SetStateAction<boolean>>;
 };
 
+export type RegCodeItem = {
+	code: string;
+	name: string;
+	selected?: boolean;
+};
+
+export type ChipType = {
+	dateChip: string;
+	distChips: RegCodeItem[];
+};
+
 export default {};
